feat: nickname 필드를 Member 엔티티에서 Account 엔티티로 이동
