+++ backend/turip-app/src/main/java/turip/account/domain/Account.java
@@ -31,12 +31,24 @@ public class Account {
     @Column(name = "role", nullable = false)
     private Role role;
 
+    @Column(name = "nickname", unique = true, nullable = false)
+    private String nickname;
+
+    public Account(Role role, String nickname) {
+        this.role = role;
+        this.nickname = nickname;
+    }
+
     public Account(Role role) {
         this.role = role;
     }
 
     public static Account createUserAccount() {
         return new Account(Role.USER);
+    }
+
+    public void updateNickname(String nickname) {
+        this.nickname = nickname;
     }
 
     public boolean isAdmin() {
