+++ backend/turip-app/src/main/java/turip/account/service/AccountService.java
@@ -1,11 +1,13 @@
 package turip.account.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import turip.account.domain.Account;
 import turip.account.repository.AccountRepository;
 import turip.common.exception.ErrorTag;
+import turip.common.exception.custom.InternalServerException;
 import turip.common.exception.custom.NotFoundException;
 import turip.favorite.repository.FavoriteContentRepository;
 import turip.favorite.service.FavoriteFolderAccountService;
@@ -16,15 +18,20 @@ import turip.favorite.service.FavoriteFolderService;
 public class AccountService {
 
     private final AccountRepository accountRepository;
+    private final AccountCreateService accountCreateService;
     private final FavoriteContentRepository favoriteContentRepository;
     private final FavoriteFolderService favoriteFolderService;
     private final FavoriteFolderAccountService favoriteFolderAccountService;
 
     @Transactional
     public Account create() {
-        Account savedAccount = accountRepository.save(Account.createUserAccount());
-        favoriteFolderService.createDefaultFavoriteFolder(savedAccount);
-        return savedAccount;
+        try {
+            Account savedAccount = accountCreateService.createUserAccount();
+            favoriteFolderService.createDefaultFavoriteFolder(savedAccount);
+            return savedAccount;
+        } catch (DuplicateKeyException e) {
+            throw new InternalServerException(ErrorTag.NICKNAME_CREATION_ERROR);
+        }
     }
 
     public Account getById(Long accountId) {
