+++ backend/turip-app/src/main/java/turip/account/service/MemberService.java
@@ -1,7 +1,6 @@
 package turip.account.service;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import turip.account.domain.Account;
@@ -12,7 +11,6 @@ import turip.auth.service.RefreshTokenService;
 import turip.common.exception.ErrorTag;
 import turip.common.exception.custom.BadRequestException;
 import turip.common.exception.custom.IllegalArgumentException;
-import turip.common.exception.custom.InternalServerException;
 import turip.common.exception.custom.NotFoundException;
 import turip.favorite.repository.FavoriteContentRepository;
 import turip.favorite.service.FavoriteFolderAccountService;
@@ -27,15 +25,13 @@ public class MemberService {
     private final AccountService accountService;
     private final RefreshTokenService refreshTokenService;
     private final FavoriteFolderAccountService favoriteFolderAccountService;
-    private final MemberCreateService memberCreateService;
 
     @Transactional
     public Member create(String email) {
-        Account account = accountService.create();
         try {
-            return memberCreateService.save(account, email);
-        } catch (DuplicateKeyException e) {
-            throw new InternalServerException(ErrorTag.NICKNAME_CREATION_ERROR);
+            Account account = accountService.create();
+            Member member = new Member(account, email, true);
+            return memberRepository.save(member);
         } catch (IllegalArgumentException e) {
             throw new BadRequestException(e.getErrorTag());
         }
